# product-api/swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
    Product:
        properties:
            description:
                description: the description for this product
                type: string
                x-go-name: Description
            id:
                description: the id for this product
                format: int64
                minimum: 1
                type: integer
                x-go-name: ID
            name:
                description: the name for this product
                type: string
                x-go-name: Name
            price:
                description: the price for this product
                format: float
                minimum: 0.01
                type: number
                x-go-name: Price
            sku:
                description: the SKU for this product
                pattern: '[a-z]+-[a-z]+-[a-z]+'
                type: string
                x-go-name: SKU
        required:
            - id
            - name
            - price
            - sku
        type: object
        x-go-package: github.com/peeush-agarwal/learn-golang/product-api/data
info:
    description: Documentation for Product API
    title: Product API
    version: 1.0.0
paths:
    /products:
        get:
            description: Returns a list of products
            operationId: listProducts
            responses:
                "200":
                    $ref: '#/responses/productsResponse'
            tags:
                - products
        post:
            description: Add a new product
            operationId: addProduct
            parameters:
                - description: |-
                    Product data structure to create or update
                    Note: the id field is ignored by update and create operations
                  in: body
                  name: Body
                  required: true
                  schema:
                    $ref: '#/definitions/Product'
            responses:
                "201":
                    $ref: '#/responses/noContent'
            tags:
                - products
    /products/{id}:
        delete:
            description: Delete the specified product
            operationId: deleteProduct
            parameters:
                - description: The id of the product to delete from the database
                  format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "201":
                    $ref: '#/responses/noContentResponse'
            tags:
                - products
produces:
    - application/json
responses:
    noContentResponse:
        description: No content is returned by this API endpoint
    productsResponse:
        description: A list of products returned in the response
        schema:
            items:
                $ref: '#/definitions/Product'
            type: array
schemes:
    - http
swagger: "2.0"